Rename database package to match conventions.
