+++ db/mongo.go
@@ -1,4 +1,4 @@
-package dataBase
+package db
 
 import (
 	"gopkg.in/mgo.v2"
