+++ handlers/userHandlers.go
@@ -3,11 +3,11 @@ package handlers
 import (
 	"net/http"
 	"encoding/json"
-	"github.com/DVI-GI-2017/Jira__backend/dataBase"
+	"github.com/DVI-GI-2017/Jira__backend/db"
 )
 
 func Test(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(200)
-	json.NewEncoder(w).Encode(dataBase.UsersListFromFakeDB)
+	json.NewEncoder(w).Encode(db.UsersListFromFakeDB)
 }
